+++ javalib/src/main/scala/java/lang/Math.scala
@@ -1,6 +1,6 @@
 package java.lang
 
-import scalanative.runtime.Intrinsics._
+import scalanative.runtime.LLVMIntrinsics._
 import scalanative.libc.{math => cmath}
 
 object Math {
