+++ nativelib/src/main/scala/scala/scalanative/runtime/LLVMIntrinsics.scala
@@ -7,7 +7,7 @@ import native._
  * @see [[http://llvm.org/releases/3.7.0/docs/LangRef.html#intrinsic-functions LLVM intrinsics functions]]
  */
 @extern
-object Intrinsics {
+object LLVMIntrinsics {
   @struct class IntOverflow(val value: Int, val flag: Boolean)
   @struct class LongOverflow(val value: Long, val flag: Boolean)
